+++ WorldGenerationProject/src/main/java/com/mycompany/worldgenerationproject/GUI/Math/Rect.java
@@ -12,10 +12,10 @@ package com.mycompany.worldgenerationproject.GUI.Math;
 public class Rect
 {
     
-    private float xMin;
-    private float xMax;
-    private float yMin;
-    private float yMax;
+    private double xMin;
+    private double xMax;
+    private double yMin;
+    private double yMax;
     
     
     
@@ -30,7 +30,7 @@ public class Rect
         yMax = 0;
     }
     
-    public Rect(float xMin_, float xMax_, float yMin_, float yMax_)
+    public Rect(double xMin_, double xMax_, double yMin_, double yMax_)
     {
         xMin = xMin_;
         xMax = xMax_;
@@ -43,42 +43,42 @@ public class Rect
     
     
     
-    public float getxMin()
+    public double getxMin()
     {
         return xMin;
     }
     
-    public void setxMin(float xMin)
+    public void setxMin(double xMin)
     {
         this.xMin = xMin;
     }
     
-    public float getxMax()
+    public double getxMax()
     {
         return xMax;
     }
     
-    public void setxMax(float xMax)
+    public void setxMax(double xMax)
     {
         this.xMax = xMax;
     }
     
-    public float getyMin()
+    public double getyMin()
     {
         return yMin;
     }
     
-    public void setyMin(float yMin)
+    public void setyMin(double yMin)
     {
         this.yMin = yMin;
     }
     
-    public float getyMax()
+    public double getyMax()
     {
         return yMax;
     }
     
-    public void setyMax(float yMax)
+    public void setyMax(double yMax)
     {
         this.yMax = yMax;
     }
